requirements.txt
streamlit
pandas
plotly
matplotlib
pillow
bcrypt
openai
